OCSADV-272 Move regression tests to gems package #time 5m
