+++ bundle/edu.gemini.ags/src/test/scala/edu/gemini/ags/gems/UCAC3RegressionSpec1.scala
@@ -1,10 +1,11 @@
-package edu.gemini.ags.impl
+package edu.gemini.ags.gems
 
-import edu.gemini.ags.gems._
 import edu.gemini.catalog.api._
-import edu.gemini.spModel.core._
+import edu.gemini.shared.util.immutable.ScalaConverters._
+import edu.gemini.shared.util.immutable.{None => JNone}
+import edu.gemini.spModel.core.AngleSyntax._
 import edu.gemini.spModel.core.Target.SiderealTarget
-
+import edu.gemini.spModel.core._
 import edu.gemini.spModel.gemini.gems.Canopus
 import edu.gemini.spModel.gemini.gsaoi.GsaoiOdgw
 import edu.gemini.spModel.gemini.obscomp.SPSiteQuality
@@ -13,10 +14,6 @@ import edu.gemini.spModel.gems.GemsGuideStarType
 import edu.gemini.spModel.target.SPTarget
 import edu.gemini.spModel.target.env.GuideGroup
 import org.specs2.mutable.Specification
-import edu.gemini.spModel.core.AngleSyntax._
-
-import edu.gemini.shared.util.immutable.ScalaConverters._
-import edu.gemini.shared.util.immutable.{None => JNone}
 
 /**
  * Regression test on target TYC 8345-1155-1, with UCAC3 data and nominal conditions
