+++ bundle/edu.gemini.ags/src/test/scala/edu/gemini/ags/gems/mascot/MascotGuideStarSpec.scala
@@ -1,6 +1,6 @@
 package edu.gemini.ags.gems.mascot
 
-import edu.gemini.ags.impl.UCAC3Regression
+import edu.gemini.ags.gems.UCAC3Regression
 import edu.gemini.catalog.api.{ppmxl, RadiusConstraint, CatalogQuery}
 import edu.gemini.catalog.votable.{TestVoTableBackend, VoTableClient}
 import edu.gemini.spModel.core.Target.SiderealTarget
